Complete swap to big int for storage. 10.41 false positive, 287 kb memory usage.

+++ ds-for-storage-and-retrieval/00-bloom-filter/hashing.go
@@ -8,7 +8,7 @@ import (
 
 var hashFn = fnv.New64a()
 
-func Hashing(s string, index int, size int) uint {
+func Hashing(s string, index int, size int64) uint {
 	write := []byte(s + fmt.Sprint(index))
 	hashFn.Write(write)
 	usize := uint64(size)
